DL-Potato plant disease prediction/api/requirements.txt
tensorflow==2.9.1
fastapi
uvicorn
python-multipart
pillow==9.2.0
tensorflow-serving-api==2.9.1
matplotlib==3.5.2
numpy==1.23.1
